Delete routes map and file. Move review route to server.js
